[FEAT] filteredTalents 유틸 개발 및 적용 | Make and Apply filteredTalents TALEARNT-373
