+++ src/hook/useFilteredTalents.ts
@@ -1,19 +1,6 @@
 import { useMemo } from "react";
 
-import { CATEGORIZED_TALENTS_LIST } from "@common/common.constants";
+import { filteredTalents } from "@utils/filteredTalents";
 
 export const useFilteredTalents = (talentCodes: number[]) =>
-  useMemo(() => {
-    const result: { talentCode: number; talentName: string }[] = [];
-    const talentCodeSet = new Set(talentCodes);
-
-    for (const { talents } of CATEGORIZED_TALENTS_LIST) {
-      for (const talent of talents) {
-        if (talentCodeSet.has(talent.talentCode)) {
-          result.push(talent);
-        }
-      }
-    }
-
-    return result;
-  }, [talentCodes]);
+  useMemo(() => filteredTalents(talentCodes), [talentCodes]);
